pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "db-backup-utility"
version = "0.1.0"
description = "Command-line database backup and restore utility scaffold"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "APScheduler>=3.10,<4",
  "python-dotenv>=1,<2",
  "requests>=2,<3",
  "boto3>=1.34",
]

[project.optional-dependencies]
dev = ["pytest>=8,<9"]
cloud = [
  "google-cloud-storage>=2.16",
  "azure-storage-blob>=12.19",
]
databases = [
  "pymysql>=1.1",
  "psycopg2-binary>=2.9",
  "pymongo>=4.6",
]

[project.scripts]
dbbackup = "db_backup_utility.cli.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
